pyproject.toml
[tool.poetry]
name = "primary2024-dashboard"
version = "0.1.0"
description = ""
authors = ["John R. Eakin <johnreakin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
streamlit = "^1.31.1"
snowflake-snowpark-python = "^1.12.1"
python-dotenv = "^1.0.1"
matplotlib = "^3.8.2"
pydantic = "^2.6.1"
tabulate = "^0.9.0"
sqlalchemy = ">=1.4.0,<2.0.0"
psycopg2 = "^2.9.9"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.0"
urllib3 = "^2.2.0"
js2py = "^0.74"
cfscrape = {path = "/Users/johneakin/cloudflare-scrape"}
nameparser = "^1.1.3"
tomli = "^2.0.1"
snowflake-connector-python = "^3.7.0"
snowflake-sqlalchemy = "^1.5.1"
tqdm = "^4.66.2"
rich = "^13.7.0"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
fastui = "^0.5.2"
httpx = "^0.27.0"
selenium = "^4.18.1"
bs4 = "^0.0.2"
aiofiles = "^23.2.1"
watchdog = "^4.0.0"
pydantic-extra-types = "^2.6.0"
phonenumbers = "^8.13.31"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
